+++ Homeworks/Homework3/WaterCard.h
@@ -18,9 +18,9 @@ public:
     WaterCard(MyString, size_t, size_t);
 
     void changeCardName(MyString);
-    void changeBaseDemage(int);
-    void changeBonusDemage(int);
-    void addPower(size_t);
+    void changeBaseDemage(size_t);
+    void changeBonusDemage(size_t);
+    void addPower(int);
 
     MyString seeCardName() const;
     size_t BaseDemage() const;
